Do not close Filesystem objects
The hbase bulkload job was breaking because I was closing the "same"
Filesystem object as the one HBase used to write the table partitions.
Since hbase sets the partition file as "deleteOnExit", it was getting
deleted immediately after it was created.

I dislike not closing objects that implement Closable, but I guess
Hadoop has its own way to do things.